create TaskCalendarAddressDto

+++ src/main/java/net/focik/addresses/infrastructure/jpa/IAddressDtoRepository.java
@@ -3,5 +3,5 @@ package net.focik.addresses.infrastructure.jpa;
 import net.focik.addresses.infrastructure.dto.AddressDbDto;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-interface IAddressDtoRepository extends JpaRepository<AddressDbDto, Integer> {
+interface IAddressDtoRepository extends JpaRepository<AddressDbDto, Long> {
 }
